chore: refactor foreign call executors
