+++ tooling/nargo/src/ops/execute.rs
@@ -10,9 +10,8 @@ use acvm::{acir::circuit::Circuit, acir::native_types::WitnessMap};
 use acvm::{AcirField, BlackBoxFunctionSolver};
 
 use crate::errors::ExecutionError;
+use crate::foreign_calls::ForeignCallExecutor;
 use crate::NargoError;
-
-use super::foreign_calls::ForeignCallExecutor;
 
 struct ProgramExecutor<'a, F, B: BlackBoxFunctionSolver<F>, E: ForeignCallExecutor<F>> {
     functions: &'a [Circuit<F>],
