+++ tooling/nargo/src/ops/mod.rs
@@ -4,7 +4,6 @@ pub use self::compile::{
     compile_workspace, report_errors,
 };
 pub use self::execute::{execute_program, execute_program_with_profiling};
-pub use self::foreign_calls::{DefaultForeignCallExecutor, ForeignCall, ForeignCallExecutor};
 pub use self::optimize::{optimize_contract, optimize_program};
 pub use self::transform::{transform_contract, transform_program};
 
@@ -13,7 +12,6 @@ pub use self::test::{run_test, TestStatus};
 mod check;
 mod compile;
 mod execute;
-mod foreign_calls;
 mod optimize;
 mod test;
 mod transform;
