+++ tooling/nargo_cli/benches/criterion.rs
@@ -115,7 +115,7 @@ fn criterion_test_execution(c: &mut Criterion, test_program_dir: &Path, force_br
     let artifacts = RefCell::new(None);
 
     let mut foreign_call_executor =
-        nargo::ops::DefaultForeignCallExecutor::new(false, None, None, None);
+        nargo::foreign_calls::DefaultForeignCallExecutor::new(false, None, None, None);
 
     c.bench_function(&benchmark_name, |b| {
         b.iter_batched(
